ParamDefs: The Grand Reshuffling of Parameters
Finally we get to move the parameters around, first of all to
get the pages in a logical order, and secondarily to get the
parameters in a logical order on the individual pages.

Also tidy up the naming and line breaks (line breaks on Zynthian
can occur when there are spaces in the label strings).

Both envelopes now have the same Attack, Decay etc labels.