MiMi-d: Use integer values for enumerated parameters
Instead of using a range of 0.0 .. 1.0 for enumerated parameters,
with the increment depending on the number of values, use integer
values, starting at 0 (and an increment of 1).

An advantage of this is that it is possible to upgrade existing
enumerated parameters by adding to the list of values, which
will still be backwards compatible with existing saved programs.