ParamDefs: Use SP_MAX for max value of enum parameters
To avoid having to manually count and specify the number of
enumerated values for the MAX value, use the special value SP_MAX
instead. It may only be used for the MAX value (well, technically
it could be used for MIN as well, but not usefully), and not the
DEFAULT value, since the DEFAULT value is used in a different
context.

Also set the default voice count to 6 instead of 8 in honor of
the MiMi-a.