AdssrEnvelope: Fix division-by-zero failure when going into decay (#5)
If the default values (ADSR mode on, Sustain level zero) are first
loaded, and then a patch setting ADSR mode, depending on which
order the parameters are set, this can result in a
division-by-zero in calculate_sustain_asymptote().

Fix this by recalculating sustain_asymptote before every
usage.

(The envelope code should really be rewritten to avoid
recalculating coef at every phase transition; coefs should
be set when the parameters are set, separately for
each envelope stage).