Voice: Bugfix: avoid click at start of sound in economy mode
Avoid click at start of sound when fenv modulation is turned up,
caused by the fenvd delay line being reset to zero, but actually
containing bipolar envelope values, so the envelope reset state
value is actually -1 instead of 0. The result was a slight hump
in the audio output for each note played after restarting the
voice in economy mode.

Fix this by moving the delay so it delays the envelope prior
to converting to bipolar.