MiMi-d: Add support for PARAMHINTS, for integer parameters (#9)
PARAMHINTS allows continuous parameters to have hints, primarily
intended for kParameterIsInteger which gives parameters
integer-only values if properly honoured by the host (Zynthian
does).

Use this for oscillator pitch, bend range, voice count and
global octaves. Note that the old parameters all started at
a value of 0, regardless of the actual displayed lowest
value, but we now set the actual range to be used, which affects
the voice count and octaves ranges.

Put the hints in the high half of a uint32_t, keeping the enum
count in the lower half. This way the return value from
setParameterEnumerationValues() can contain both an enum count
and a hint (in the case of boolean parameters).

Additionally adjust the defaults to be more sane for some
of these parameters.