Voice: Adjust filter offset to get range 20 .. 20 kHz
Instead of max filter cutoff (without modulation) being around 33 kHz,
change to around 20 kHz, which also nicely reduces the low end from
34 to 20 Hz, useful above all when the filter pole count is low.

At a cutoff setting of 5, the resulting frequency will then be 622 Hz,
which is a cent away from Eb5.

In conjuction with this, change the base note of filter tracking from
F to Eb (midi note 39 = 81-42), so that at the zero tracking point,
the filter pitch will be Eb.

As before, the cutoff control represents one unit per octave, so that
a setting of 6 is one octave above 5 for instance.