Voice: Square LFO amount to get better resolution
Square the initial amount of LFO modulation to get better resolution.
Since the modulation amount is in the interval 0..1, the total range
will not change, only the scaling.

Note that this only affects the initial value, not the controller
amount or the amount of modulation that the selected controller
adds.