Voice: Adjust oscillator pitches down one octave
On the MiMi-a, an oscillator pitch of 24 results in a middle C
(261.63 Hz) when MIDI note 60 (middle C) is played. Adjust
oscillator pitch calculation to get the same.

We also need to make a complementary change in the filter tracking
calculations to avoid the filter tracking changing as a result
of this.