Add and use new SRandom class to replace Juce Random class
SRandom is based on the same underlying LCG random generator
as the Juce random number generator, itself based on a well-known
algorithm,, but simplifying operations to use only 32 bit floats.
This means that the lowest bits will not be particularly random,
but this is deemed inconsequential in the MiMi-d application.