MiMi-d, Lfo: Enable support for bpm synced LFOs
In tempo sync mode, have 11 modes, to line up nicely with the
0..10 parameter range (11 values 0, 1, 2, 3, 4, 5, 6, 7, 8, 9,
10, 11), selecting from a set of hopefully useful sync values.

Save the bpm value when set so that the phase increment (= frequency)
can be recalculated if the rate (sync ratio) is changed, even if the
transport is stopped.

Rewrite sync mode rate using table rather than switch statement.
Also, if parameter is out of range, leave ratio unchanged.

Whan tempo syncing, beatpos includes the BPM (i.e. it increases
at the BPM rate), so when calculating the synced phase, do not
multiply by phaseInc, which already has a BPM factor, but just
by the syncRatio.

The result is otherwise that the resulting rate is (BPM/60)^2
resulting in a correctly synced rate at 60 BPM, but twice as
fast (2^2) at 120 BPM.

Update beats per minute as soon as the bpm value is updated,
rather than waiting for playing to be true, so that the bpm is
updated even when the transport is stopped.