AdssrEnvelope: Add hold phase (#7)
Add steady-state hold phase between attack and decay.

In contrast to the slope parameters, whose minimum value is 1 ms
to mimic the way an analog envelope generator operates, the minimum
hold time is set to 10 us. While this is (usually) less than an
(output) sample period, it does guarantee that there is absolutely
no hold period when the hold time parameter is set low enough, and
due to the scaling of time parameters, it doesn't affect the
distribution of higher values in any significant way.

Note that this commit only adds the hold phase to the envelope,
plus the parameter setting callback. The actual hold feature
has been deemed out of scope for the MiMi-d at this time, because
it requires two parameter pages to reach the note-on-to-off
envelope parameters. It's left in the code if someone wants to
add the feature in a private branch, or for a future change.