AdssrEnvelope: Add hold phase (#7)
Add steady-state hold phase between attack and decay.

In contrast to the slope parameters, whose minimum value is 1 ms
to mimic the way an analog envelope generator operates, the minimum
hold time is set to 10 us. While this is (usually) less than an
(output) sample period, it does guarantee that there is absolutely
no hold period when the hold time parameter is set low enough, and
due to the scaling of time parameters, it doesn't affect the
distribution of higher values in any significant way.