Add envelope hold parameter (#7)
Add envelope hold parameter to parameter list and also add
associated setter function.

In contrast to the slope parameters, whose minimum value is 1 ms
to mimic the way an analog envelope generator operates, the minimum
hold time is set to 10 us. While this is (usually) less than an
(output) sample period, it does guarantee that there is absolutely
no hold period when the hold time parameter is set low enough, and
due to the scaling of time parameters, it doesn't affect the
distribution of higher values in any significant way.