DelayLine: Use template parameter for delay line dimension
Use template parameter DM for delay line dimensioning, rather than
having a DEMAX maximum value. Apart from reducing the size of the
delay buffer which is an advantage from a cache point of view,
it also means we can remove the addition and masking with DM-1,
which helps realtime performance.

The only downside is that DM now needs to be a power of two, which
in practice it is anyway.