Oscs, BlepData: Restructure BLEP/BLAMP tables for simpler access
Change the contents of the BLEP and BLAMP tables so that only a
single loop is required to add them in, and that the sign of
the mixing is always positive.

For the BLEP tables, this means changing the sign of the first
half of the table. This avoids having two loops to mix in the impulse
center.

For the BLAMP tables, the sign of the whole table is changed. There
is already a single loop to add in the BLAMP.

Also take the opportunity to reformat the tables to something a bit
easier to handle in an ordinary editor. Since the data is oversampled
by a factor of 64, group the table entries into groups of 64.

NOTE: This fails because for the highest phase of the BLEP, the adjacent
BLEP which is used for the interpolation is the same as the lowest
phase, albeit shifted one sample. This means that the change of sign
happens one sample off, resulting in random clicking in the resulting
composite waveform.