Oscs, BlepData: Restructure BLEP/BLAMP tables for faster access
In order to increase the locality of the BLEP/BLAMP tables, restructure
the waveforms so that the subsampled entries are adjacent to each other
rather than being separated by B_OVERSAMPLING. The downside is that
the adjacent oversampled sample in the original table is now
Bleplen (Samples * 2) entries away. Still, this means that the region
to be accessed for a single BLEP/BLAMP addition is 2 * 2 * Samples
= 64 samples rather than 64 samples short of 2049 entries with the
original tables.

Tests however on Zynthian w/ RPi4 reveal that there is no discernable
difference between this and the original implementation. Perhaps randomness
of where the table accesses actually hit mean that basically the whole
table must be loaded regardless of their organization, and furthermore
the table is actually bigger with this version.

However, in order to change the sign in the table data so that single
loops can be used to add in the BLEPs (in another commit), there must
be a separate table for the final phase in the oversampled BLEP, to
avoid an incorrect interpolation when one phase is interpolated with
the next higher one, and the next higher one is actually the first
phase, shifted one sample.