Presets: Add Default.ttl
Add new Default.ttl which has all parameters (including new hold
and filterhold) set to their default values. This can be used to
create a clean slate to start from (note that the synth will be
completely silent though, as all volumes are set to 0 and the
oscillator waveforms turned off), or as a preset to load before
loading presets saved with older versions of the plugin not
having all parameters, to avoid the values of the newer
parameters leaking through from whatever settings they had prior
to loading the older preset.