Makefile: Add fixup target for lv2:designation param feature
Add fixup target which post-processes the MiMi-d_dsp.ttl file
extracted from the lv2 .so file, to add lv2:designation param
for the envelope parameters.

Thus ultimately allows Zynthian to draw a graphic rendering
of the envelope shape.

For this particular commit, this function is disabled, to be
enabled on the zynthian branch.