Makefile: Add fixup target for lv2:designation param feature
Add fixup target which post-processes the MiMi-d_dsp.ttl file
extracted from the lv2 .so file, to add lv2:designation param
for the envelope parameters.

Thus ultimately allows Zynthian to draw a graphic rendering
of the envelope shape.

Currently this is disabled, because Zynthian requires all
envelope parameters to be in the same page group, which
MiMi-d cannot satisfy, since the release parameters for
both envelopes are on the same page.