ParamDefs: Utilize one param group for each envelope
Zynthian requires that all envelope parameters be in the
same parameter group in order for the graphic renderer
to properly display all envelope parameters regardless
of which envelope page is current.

This has the unfortunate side effect that the release
parameters end up on single parameter pages, increasing
the number of pages needed to reach the envelope parameters.