Voice: Only ramp down loudness env when resetting
When a voice is retriggered with envelope reset enabled, it is only
necessary to smoothly ramp down the loudness envelope, as it is in
the VCA that the most severe clicks would occur when restarting the
voice. Sure, changes to the filter can also be abrupt, but that
could be due to other reasons that the voice retrigger (new pitch
with keyfollow advanced, or a different or restarting LFO).

Doing this means we can remove the fenvd delay line and
just delay the filter cutoff.