VariSawOsc, Oscillators, Voice: Rework VariSaw slope calculation
Rework slope calculation for VariSaw so that the gradient is
a factor, with 1 representing a standard sawtooth, higher values
corresponding to sharper slopes, using exponential scaling,
in a similar vain to MiMi-a shape/slope parameter.

Values < 1 result in a flatter and flatter sawtooth, also like
the MiMi-a.

Set gradient limit to correspond to 58 microseconds (empirically
determined and comparing with MiMi-a), and a derating factor
at 0.1 above that, so that the shape parameter just doesn't hit
a brick wall above a certain value. This sortof mimics the
peak current limitation in a real world exponential generator
(as used in the MiMi-a) causing the generator to go flat at
high output currents.