FastExp: Add (superfast) exp function for saw shape parameter
And scale it so that every whole step of input parameter change
corresponds to an approximate doubling.

For speed, we don't use the correction table, because the shape
parameter doesn't require the precision.

Also, by setting the fudge term appropriately, we get minimum deviation
from a true exponential at x = 0 (and each "octave" above that);
in fact, we now get superfast_exp(0) == 1.0f . This is at the expense
of the maximum deviation which now instead of being +/- 50 cents
(with the fudge term being set as for the pitch exponentiator)
(roughly) is +100 / -0 cents . For the shape this is not as
critical as it would be for a pitch exponential.

In the end it means that a shape setting of 0.0f will result in
an exact gradient of 1.0f, which can be used to distinguish a
true saw from a variable slope saw.