Oscillators: Localize waveform progression
Instead of having separate switch statements for the antialising
and waveform generation, put them in a single one, while
putting the phase reset code in between the process and getValue
calls for each one. The resulting code will be longer, but avoids
a branch, and should be easier for the compiler to optimize.