Oscillators, VariSaw: Revert to sawtooth when gradient == 1
When gradient is exactly 1, revert to pure sawtooth wave. This
keeps the CPU load down for backwards compatibility.

Since VariSaw is now not called with breakpoint == 1.0, it would
seem an interesting option not to have to consider this case
specifically, when going through the various oscillator states.
We could then cut a couple of branches, and perhaps also bunch the
blamp and blep insertions into a common function for those cases
where both are needed. Howwever, this is actually not true: Although
now never called when gradient == 1.0, the gradient can still be
> 1.0f, at which point breakpoint remains at == 1.0. The resulting
output will be an amplitude reduced sawtooth, which can not be handled
by the ordinary saw oscillator. Pity. Well, at least we can stop
worrying about optimizing this case...