Presets: Reset pulsewidth to zero for non pulse waves
When an oscillator waveform is not set to pulse, there is no point
in having the pulsewidth advanced past 0, so reset it in this case.

Strangely, although this has been done previously, this has crept
in again, possibly during some preset updates phase.

Ultimately, this is in preparation for the pulse width becoming
a generalized shape parameter also applicable to the triangle
and sawtooth waveforms.