Motherboard, Voice: Update voice aftertouch smoother unconditionally
This avoids the aftertouch getting an unexpected initial value
if poly aftertouch is active after voices has stopped playing.

This is in part a bit of an esoteric case: Normally, the (poly)
aftertouch value will naturally drop to 0 before note off. But in
the unlikely event that the note off arrives first, we want the
aftertouch to drop off smoothly rather than just freeze.

A more realistic case is channel aftertouch: since the aftertouch
smoother is now per-voice, we need to feed it with updated aftertouch
values so that it is correctly poised when it gets gated on in the
future as a result of a Note On.