SynthEngine, VoiceAllocator: Handle poly AT of restored notes
When notes are restored, i.e. allocated to a new voice as the
result of another note being released, freeing up a voice, we
need to set the aftertouch value for the voice to the appropriate
value for the note in question (i.e., the current aftertouch value
for that note). Therefore, the voice allocator keeps track of the
aftertouch values for each note (similar to the way it keeps track
of the velocity for each played note, for the same reason), and
when a new voice is allocated, in the poly aftertouch case,
the aftertouch value for the newly allocated voice is set
accordingly when the voice is gated on.

For the channel aftertouch case, this is not necessary, as when
receiving a channel aftertouch message, all voices are updated
with the new aftertouch value, so there is no need to do it
in the voice allocator.