Voice, SynthEngine, Oscillators: Rename (most) *pw* variables to *sh*
Since what was the pulse width parameter now affects the shape of
all oscillator waveforms, change the names of the corresponding
variables as well.

Note that in some cases 'pw' really does stand for 'pulse width'
(i.e. in reference to the pulse wave oscillator), and in other
cases 'pw' stands for 'pitch wheel' so don't change those.