Lfo: Add setShape callback
To be used when the waveshape is set using a continuous parameter
SAW - TRI - RSAW - PUL - SQU -RPUL
i.e. the left half of the parameter range are trapezoid waveforms
and the right half of the parameter range are pulse waveforms.

Of special note is that <= 0.5 is used for the comparison for
selecting the trapezoid vs square waveform.  These means that if
the shape parameter is set to 5, it will result in a reverse
sawtooth (i.e., rising), and in particular, when used in oneshot
mode, it will behave as an AR envelope, i.e. start low, then
rise up and stop there, just as LFO1 and 2 do when the waveform
is RSaw, as opposed to all other waveform settings when it behaves
as an AD envelope, i.e. runs the whole LFO cycle through back to
zero before stopping.